+++ src/main/java/at/technikum/springrestbackend/controller/AuthController.java
@@ -1,7 +1,7 @@
 package at.technikum.springrestbackend.controller;
 
-import at.technikum.springrestbackend.dto.TokenRequest;
-import at.technikum.springrestbackend.dto.TokenResponse;
+import at.technikum.springrestbackend.dto.auth.TokenRequest;
+import at.technikum.springrestbackend.dto.auth.TokenResponse;
 import at.technikum.springrestbackend.service.AuthService;
 import jakarta.validation.Valid;
 import org.springframework.web.bind.annotation.PostMapping;
