+++ src/main/java/at/technikum/springrestbackend/dto/lawyer/CreateLawyerRequest.java
@@ -1,4 +1,4 @@
-package at.technikum.springrestbackend.dto;
+package at.technikum.springrestbackend.dto.lawyer;
 
 import at.technikum.springrestbackend.model.Specialization;
 import at.technikum.springrestbackend.model.Lawyer;
