+++ src/main/java/at/technikum/springrestbackend/dto/lawyer/GetLawyerProfilesRequest.java
@@ -1,4 +1,4 @@
-package at.technikum.springrestbackend.dto;
+package at.technikum.springrestbackend.dto.lawyer;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
